fix: change db options and attributes
Previously the database options and attributes were provided as
PDO key to corresponding value; usually string or another PDO
constant. Unfortunately the PDO constants are integers which
makes them vulnerable to PHP array_merge ambiguisation. To make
things worse the way merges happen can be extremely confusing,
for example if an array has 1002 (a potential PDO constant value)
new values may be coerced into 1003, 1004, etc, instead of their
actual values which might something 1007. Obviously this
corruption is very annoying and hard to spot; unless you
individually check the constant values.

To avoid this nonsense tacking hold further; the configuration
system for options and attributes has been changed to an
associative array that is immune to it's destructive effects.

In addition, names are required for configurations, and should
the keys be associated with null they will be ignored. To avoid
backwards compatibility issues non-array values are also ignored.